Fixed year picker in my vacation reports. Also moved FileController to drive namespace.
