+++ Presentation.Web/App/Vacation/MyVacationReports/View.html
@@ -8,7 +8,12 @@
 <div class="container-fluid">
     <div class="row">
         <div class="col-xs-4 col-lg-3 to-front">
-            <input kendo-datepicker culture="'da-DK'" k-options="mvrCtrl.dateOptions" ng-model="mvrCtrol.vacationYearString" k-ng-model="mvrCtrl.vacationYear" k-on-change="mvrCtrl.refreshGrid()" />
+            <select kendo-drop-down-list
+                    k-data-text-field="'YearString'"
+                    k-data-value-field="'Year'"
+                    k-data-source="mvrCtrl.vacationYears"
+                    ng-model="mvrCtrl.vacationYear"
+                    k-on-change="mvrCtrl.refreshGrid()"></select>
         </div>
     </div>
     <div kendo-grid="mvrCtrl.vacationReportsGrid" class="margin-top-10 scrollable" options="mvrCtrl.vacationReportsOptions"></div>
