+++ Presentation.Web/Controllers/Drive/FileController.cs
@@ -1,27 +1,18 @@
 ﻿using System;
-using System.Configuration;
-using System.Linq;
 using System.Net;
 using System.Web.Http;
-using System.Web.Http.Controllers;
-using System.Web.Mvc;
-using System.Web.Optimization;
-using Core.ApplicationServices;
 using Core.ApplicationServices.FileGenerator;
 using Core.ApplicationServices.Logger;
 using Core.DomainModel;
 using Core.DomainServices;
-using Ninject;
 
-namespace OS2Indberetning.Controllers
+namespace OS2Indberetning.Controllers.Drive
 {
     public class FileController : BaseController<DriveReport>
     {
         private readonly IGenericRepository<DriveReport> _repo;
 
         private readonly ILogger _logger;
-
-
 
         public FileController(IGenericRepository<DriveReport> repo, IGenericRepository<Person> personRepo, ILogger logger) : base(repo, personRepo)
         {
